finish chains new should read defaults by default
